Move away from table
