+++ webapp/components/test-file-results.js
@@ -5,9 +5,9 @@
  */
 
 import '../node_modules/@polymer/paper-toggle-button/paper-toggle-button.js';
-import '../node_modules/@polymer/polymer/lib/elements/dom-if.js';
 import { html, PolymerElement } from '../node_modules/@polymer/polymer/polymer-element.js';
 import { LoadingState } from './loading-state.js';
+import './test-file-results-table.js';
 import { TestRunsUIQuery } from './test-runs-query.js';
 import { TestRunsQueryLoader } from './test-runs.js';
 import './wpt-colors.js';
@@ -45,19 +45,11 @@ class TestFileResults extends WPTFlags(LoadingState(TestRunsUIQuery(
       </paper-toggle-button>
     </div>
 
-    <template is="dom-if" if="{{!isVerbose}}">
-      <test-file-results-table-terse test-runs="[[testRuns]]"
-                                     results-table="[[resultsTable]]"
-                                     on-reftest-compare="[[onReftestCompare]]">
-      </test-file-results-table-terse>
-    </template>
-
-    <template is="dom-if" if="{{isVerbose}}">
-      <test-file-results-table-verbose test-runs="[[testRuns]]"
-                                       results-table="[[resultsTable]]"
-                                       on-reftest-compare="[[onReftestCompare]]">
-      </test-file-results-table-verbose>
-    </template>
+    <test-file-results-table test-runs="[[testRuns]]"
+                             results-table="[[resultsTable]]"
+                             on-reftest-compare="[[onReftestCompare]]"
+                             is-verbose="[[isVerbose]]">
+    </test-file-results-table>
 `;
   }
 
@@ -105,6 +97,10 @@ class TestFileResults extends WPTFlags(LoadingState(TestRunsUIQuery(
         const matches = others.filter(t => subtests.has(t.name));
         resultsTable = [first, ...matches];
       }
+    }
+    // Index needed in iron-list.
+    for (let i in resultsTable) {
+      resultsTable[i].index = i;
     }
     this.resultsTable = resultsTable;
   }
