+++ webapp/components/wpt-results.js
@@ -21,8 +21,6 @@ import './info-banner.js';
 import { LoadingState } from './loading-state.js';
 import './path-part.js';
 import { SelfNavigation } from './self-navigator.js';
-import './test-file-results-table-terse.js';
-import './test-file-results-table-verbose.js';
 import './test-file-results.js';
 import './test-results-chart.js';
 import './test-results-history-grid.js';
